# src/main/resources/templates/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger lokapi - OpenAPI 3.0
  description: ''
  termsOfService: http://swagger.io/terms/
  contact:
    email: tawfik.elyousfi@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://test.swagger.io/api/v3
...
paths:
...
  /properties:
    get:
      tags:
        - property
      summary: Get all properties
      operationId: getAllProperties
      responses:
        '200':
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
    post:
      tags:
        - property
      summary: Create a new property
      description: creation of a new property.
      operationId: createProperty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Property'
      responses:
        default:
          description: Property created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
    put:
      tags:
        - property
      summary: Update property
      description: update property existing
      operationId: updateProperty
      parameters:
        - name: propertyId
          in: path
          description: ID that need to be updated
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
    delete:
      tags:
        - property
      summary: Delete property by ID
      description: delete property .
      operationId: deleteProperty
      parameters:
        - name: propertyId
          in: path
          description: The ID that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
  /properties/{ownerId}:
    get:
      tags:
        - property
      summary: Get property by owner ID
      description: ''
      operationId: getPropertyByOwnerId
      parameters:
        - name: ownerId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Property'
  /properties/{ownerId}/addresses:
    get:
      tags:
        - property
      summary: Get property addresses by owner ID
      operationId: getAddressPropertyByOwnerId
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                    address:
                      type: string
                    city:
                      type: string
                    zip:
                      type: string
  /properties/{propertyId}:
    get:
      tags:
        - property
      summary: Get property by property Id
      description: ''
      operationId: getPropertyById
      parameters:
        - name: propertyId
          in: path
          description: 'The id that needs to be fetched. Use property1 for testing. '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          description: Invalid ID supplied
        '404':
          description: Property not found
  /properties/{propertyId}/lease:
    get:
      tags:
        - property
      summary: Get property by property Id
      description: ''
      operationId: getPropertyWithLeaseById
      parameters:
        - name: propertyId
          in: path
          description: 'Details of property with lease attached  '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          description: Invalid ID supplied
        '404':
          description: Property not found
  /leases:
    get:
      tags:
        - lease
      summary: Get all leases
      operationId: getAllLeases
      responses:
        default:
          description: List of leases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lease'
    post:
      tags:
        - lease
      summary: Create lease
      description: This can only be done by the logged in user.
      operationId: createLease
      requestBody:
        description: Created lease object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lease'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lease'
    put:
      tags:
        - lease
      summary: update lease
      description: ''
      operationId: updateLease
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lease'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lease'

    delete:
      tags:
        - lease
      summary: delete lease
      description: 'Delete add a date in delete date'
      operationId: deleteLease
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lease'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lease'
  /leases/{leaseId}:
    get:
      tags:
        - lease
      summary: Get lease by ID
      description: ''
      operationId: getLeaseById
      parameters:
        - name: leaseId
          in: path
          description: 'The ID needs to be fetched '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lease'
  /leases/{tenantId}:
    get:
      tags:
        - lease
      summary: Get lease by tenantId
      description: ''
      operationId: getLeaseByTenantId
      parameters:
        - name: tenantId
          in: path
          description: 'The tenantId needs to be fetched '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: 'successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lease'
  /leases/{ownerId}/{tenantId}:
    get:
      tags:
        - lease
      summary: Get lease by tenantId
      description: ''
      operationId: getLeaseByOwnerIdAndTenantId
      parameters:
        - name: ownerId
          in: path
          description: 'The ownerId needs to be fetched '
          required: true
          schema:
            type: integer
            format: int64
        - name: tenantId
          in: path
          description: 'The tenantId needs to be fetched '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: 'successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lease'
  /leases/{ownerId}:
    get:
      tags:
        - lease
      summary: Get lease by ownerId
      description: ''
      operationId: getLeaseByOwnerId
      parameters:
        - name: ownerId
          in: path
          description: 'The ownerId needs to be fetched '
          required: true
          schema:
            type: integer
            format: int64
      responses:
        default:
          description: 'successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lease'
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: 'theUser'
        email:
          type: string
          example: 'you@email.com'
        password:
          type: string
          example: '12345'
        roles:
          type: array
          items:
            type: string
            enum:
              - OWNER
              - TENANT
          example:
            - TENANT
            - OWNER
          description: tenant or owner status in application

        active:
            type: boolean
            example: 1
        creationDate:
          type: string
          format: date
          description: User creation date
        updateDate:
          type: string
          format: date
          description: User update date
        deleteDate:
          type: string
          format: date
          description: User detele date
      required:
        - id
        - username
        - email
        - password
        - roles
    Property:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Property ID unique
        address:
          type: string
          description: Property adress
        city:
          type: string
          description: Property city
        zip:
          type: string
          description: Property zip
        description:
          type: string
          description: Property description
        owner:
          $ref: '#/components/schemas/User'
        lease:
          $ref: '#/components/schemas/Lease'
        price:
          type: number
          format: decimal
        creationDate:
          type: string
          format: date
          description: Property creation date
        updateDate:
          type: string
          format: date
          description: Property update date
        deleteDate:
          type: string
          format: date
          description: Property detele date
      required:
        - id
        - adress
        - owner
        - city
        - zip
        - creationDate
        - updateDate
    Lease:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Lease
        property:
          $ref: '#/components/schemas/Property'
        tenant:
          $ref: '#/components/schemas/User'
        owner:
          $ref: '#/components/schemas/User'
        startDate:
          type: string
          format: date
          description: Date start of lease
        endDate:
          type: string
          format: date
          description: Date end of lease  (null if empty)
        rentAmount:
          type: number
          format: decimal
          description: rent amount
        depositAmount:
          type: number
          format: decimal
          description: desposit amount
        contract:
          type: string
          description: Lease Contrat PDF ( bail )
        creationDate:
          type: string
          format: date
          description: Lease creation date
        deleteDate:
          type: string
          format: date
          description: Lease delete date  NULL if not deleted
        updateDate:
          type: string
          format: date
          description: Lease update date can be same on creation date
        deleteDate:
          type: string
          format: date
          description: Lease delete date can be null
      required:
        - id
        - property
        - tenant
        - startDate
        - rentAmount
        - contract
        - creationDate
        - updateDate
...